// stroom-ui/src/components/DocumentEditors/FolderExplorer/types.ts
/*
 * Copyright 2018 Crown Copyright
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { ConnectDropTarget, ConnectDragSource } from "react-dnd";
import { DocRefType } from "components/DocumentEditors/useDocumentApi/types/base";

export enum DragDropTypes {
  DOC_REF_UUIDS = "docRef",
}
export interface DropCollectedProps {
  connectDropTarget: ConnectDropTarget;
  isOver: boolean;
  canDrop: boolean;
}

export interface DragCollectedProps {
  connectDragSource: ConnectDragSource;
  isDragging: boolean;
}

export interface DragObject {
  docRefs: DocRefType[];
  isCopy: boolean;
}